backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.9  # Uncomment for PostgreSQL support
playwright>=1.41.0
httpx>=0.26.0

# Third-party crawlers
bilibili-api-python>=16.2.0
beautifulsoup4>=4.12.0
litellm>=1.20.0
apscheduler>=3.10.4
fastapi-mail>=1.4.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
